refactor: image resize util
